requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
redis
python-jose
passlib[bcrypt]
pydantic